requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
celery==5.3.4
drf-yasg==1.21.7
django-environ==0.11.2
mysqlclient==2.2.0
pika==1.3.2  # For RabbitMQ integration
PyJWT==2.8.0
pillow==10.1.0
redis==5.0.1
pytest==7.4.3
pytest-django==4.7.0
